% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.exl_tbl.R
\name{print.exl_tbl}
\alias{print.exl_tbl}
\title{Prints \code{exl_tbl} objects}
\usage{
\method{print}{exl_tbl}(x, ...)
}
\arguments{
\item{x}{An \code{exl_tbl} object.}

\item{...}{Other arguments that should be past to print.}
}
\description{
This is print function for \code{exl_tbl} objects, which outputed by
\code{exlcusion_table()}. The function imporves the readability of the output.
}
