% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.exl_tbl.R
\name{print_lines}
\alias{print_lines}
\title{Funtion for printing = across the length of output}
\usage{
print_lines(i)
}
\arguments{
\item{i}{An integer vector with the number of = to print.}
}
\description{
Funtion for printing = across the length of output
}
